Don't move clang.exe to clang-<major>.exe if clang-<major>.exe already exists

+++ install-wrappers.sh
@@ -67,7 +67,7 @@ for arch in $ARCHS; do
     done
 done
 if [ -n "$EXEEXT" ]; then
-    if [ ! -L clang$EXEEXT ] && [ -f clang$EXEEXT ]; then
+    if [ ! -L clang$EXEEXT ] && [ -f clang$EXEEXT ] && [ ! -f clang-$CLANG_MAJOR$EXEEXT ]; then
         mv clang$EXEEXT clang-$CLANG_MAJOR$EXEEXT
     fi
     if [ -z "$HOST" ]; then
